Fix coding structure to fulfill SLAP principle
