+++ src/main/java/wellnus/reflection/GetCommand.java
@@ -1,7 +1,7 @@
-package seedu.duke.reflection;
+package wellnus.reflection;
 
-import seedu.duke.command.Command;
-import seedu.duke.exception.BadCommandException;
+import wellnus.command.Command;
+import wellnus.exception.BadCommandException;
 
 import java.util.ArrayList;
 import java.util.HashMap;
