+++ src/test/java/wellnus/reflection/ReturnCommandTest.java
@@ -1,7 +1,7 @@
-package seedu.duke.reflection;
+package wellnus.reflection;
 
 import org.junit.jupiter.api.Test;
-import seedu.duke.exception.BadCommandException;
+import wellnus.exception.BadCommandException;
 
 import java.util.HashMap;
