+++ src/test/java/wellnus/reflection/SelfReflectionTest.java
@@ -1,4 +1,4 @@
-package seedu.duke.reflection;
+package wellnus.reflection;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
